move short weierstrass curve test cases to instances
